+++ test/instream/cluster/stats_test.exs
@@ -1,0 +1,14 @@
+defmodule Instream.Cluster.StatsTest do
+  use ExUnit.Case, async: true
+
+  alias Instream.Cluster.Stats
+  alias Instream.TestHelpers.Connection
+
+  test "stats listing" do
+    result = Stats.show() |> Connection.execute()
+
+    %{ results: [%{ series: stats }]} = result
+
+    assert %{ name: _, tags: _, columns: _, values: _ } = hd(stats)
+  end
+end